moved main, app and style files to src and changed package scripts for vite
