+++ client/gamesRoughDraft.html
@@ -12,6 +12,9 @@
     <title>Procrastination Station</title>
   </head>
   <body>
+    <div id="root">
+
+    </div>
     <div id="hero">
         <h1> Procrastination Station </h1>
     </div>
