+++ client/package.json
@@ -2,8 +2,8 @@
     "name": "Procrastination-Station",
     "description": "description",
     "scripts": {
-        "dev": "webpack-dev-server",
-        "build": "webpack --mode production",
+        "dev": "vite",
+        "build": "vite build",
         "start": "webpack --watch",
         "vite": "vite",
         "vBuild": "vite build",
