+++ client/src/components/gameDisplayCards.jsx
@@ -1,5 +1,5 @@
-export default function Games(){
+export default function GameDisplayCards(){
     return (
         <div>Games will go here</div>
     )
-}
+};
