+++ client/src/components/gamePlay.jsx
@@ -1,0 +1,5 @@
+export default function GamePlay(){
+    return (
+        <div> Large Game playing window will go here</div>
+    )
+};
