--- client/src/pages/games.jsx
@@ -1,11 +1,0 @@
-import { Game } from '../components/gameDisplayCards'
-
-const Games = () => {
-    return (
-        <div> Home page 
-            <Game />
-        </div>
-    )
-}
-
-export default Games;
